+++ v3/crates/engine/tests/common.rs
@@ -678,6 +678,7 @@ async fn run_query_graphql_ws(
             pre_route_plugins: Vec::new(),
         }),
         metrics: graphql_ws::NoOpWebSocketMetrics,
+        handshake_headers: Arc::new(request_headers.clone()),
     };
     let (channel_sender, mut channel_receiver) =
         tokio::sync::mpsc::channel::<graphql_ws::Message>(10);