+++ v3/changelog.md
@@ -4,6 +4,10 @@
 
 ### Added
 
+- GraphQL WebSocket connections now include client headers from the initial
+  handshake request, in addition to those from the `connection_init` message.
+  These headers are forwarded to auth webhooks, plugins and data connectors.
+
 ### Fixed
 
 - Loosen restriction on `ObjectType` fields to allow them to refer to